requirements.txt
feedparser
google-generativeai
python-dotenv
beautifulsoup4
requests
scikit-learn
numpy
python-dateutil
